# swagger.yaml
openapi: 3.0.0
info:
  title: FEMCS API
  description: Fire Extinguisher Management and Compliance System
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local API Gateway
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200': { description: List of customers }
    post:
      tags: [Customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, phone]
              properties:
                fullName: { type: string }
                nationalId: { type: string }
                phone: { type: string }
                email: { type: string }
                address: { type: string }
                organizationName: { type: string }
      responses:
        '201': { description: Customer created }
...
  /inspections:
    get:
      tags: [Inspections]
      summary: List inspections
      responses:
        '200': { description: List of inspections }
    post:
      tags: [Inspections]
      summary: Record inspection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [extinguisherId, inspectorName, inspectionDate, status]
              properties:
                extinguisherId: { type: string, format: uuid }
                inspectorName: { type: string }
                inspectionDate: { type: string, format: date }
                findings: { type: string }
                status: { type: string, enum: [Passed, Requires Service, Failed] }
                nextInspectionDate: { type: string, format: date }
      responses:
        '201': { description: Inspection recorded }
  /maintenance:
    get:
      tags: [Maintenance]
      summary: List maintenance records
      responses:
        '200': { description: List of maintenance records }
    post:
      tags: [Maintenance]
      summary: Create maintenance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [extinguisherId, serviceDate, serviceCompany, technicianName]
              properties:
                extinguisherId: { type: string, format: uuid }
                serviceDate: { type: string, format: date }
                serviceCompany: { type: string }
                technicianName: { type: string }
                nextServiceDate: { type: string, format: date }
                cost: { type: number }
                description: { type: string }
                status: { type: string, enum: [scheduled, in_progress, completed, cancelled] }
      responses:
        '201': { description: Maintenance record created }
  /notifications:
    get:
      tags: [Notifications]
      summary: Get notifications for current user
      parameters:
        - in: query
          name: unread
          schema: { type: boolean }
        - in: query
          name: type
          schema: { type: string }
      responses:
        '200': { description: Notifications list with unread count }
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Marked as read }
...
  /reports/expired:
    get:
      tags: [Reports]
      summary: Expired extinguishers report
      parameters:
        - in: query
          name: format
          schema: { type: string, enum: [json, csv] }
      responses:
        '200': { description: Report data }
  /reports/expiring-soon:
    get:
      tags: [Reports]
      summary: Expiring soon report
      parameters:
        - in: query
          name: days
          schema: { type: integer, default: 90 }
        - in: query
          name: format
          schema: { type: string, enum: [json, csv] }
      responses:
        '200': { description: Report data }
  /reports/compliance:
    get:
      tags: [Reports]
      summary: Compliance report
      parameters:
        - in: query
          name: format
          schema: { type: string, enum: [json, csv] }
      responses:
        '200': { description: Compliance data }
...
  /reports/audit:
    get:
      tags: [Reports]
      summary: Audit trail (admin only)
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: action
          schema: { type: string }
      responses:
        '200': { description: Audit logs }